requirements.txt
requests>=2.31,<3
pandas>=2.2,<3
dnspython>=2.6,<3
aiohttp>=3.9,<4
async-timeout>=4.0,<5
beautifulsoup4>=4.12,<5
lxml>=5.2,<6
tinycss2>=1.3,<2
fonttools>=4.54,<5
brotli>=1.1,<2
tldextract>=5.1,<6
tqdm>=4.66,<5
# Optional: playwright>=1.46,<2